requirements.txt
dnspython~=2.8.0
requests~=2.32.5
beautifulsoup4~=4.14.3
httpx~=0.28.1
pypdf~=6.5.0
playwright~=1.57.0